// app/(content)/projects/school/infoVis-DatasetProject/components/types/pageFourType.tsx
export type PageFourType = {
    whichAvg: string;
    fixedAcidity: { label: 'Fixed Acidity', value: number };
    volatileAcidity: { label: 'Volatile Acidity', value: number };
    citricAcid: { label: 'Citric Acid', value: number };
    freeSulfurDioxide: { label: 'Free Sulfur Dioxide', value: number };
    totalSulfurDioxide: { label: 'Total Sulfur Dioxide', value: number };
    sulphates: { label: 'Sulphates', value: number };
    alcohol: { label: 'Alcohol', value: number };
};